+++ packages/react-ui/components/Modal/Modal.styles.ts
@@ -229,6 +229,12 @@ const styles = {
     `;
   },
 
+  bodyWithoutPadding() {
+    return css`
+      padding: 0 !important;
+    `;
+  },
+
   bodyAddPadding() {
     return css`
       padding-bottom: 30px;